atomic_store_n

+++ testdata/atomic/atomic.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-    int a = 3;
+    long long a = 3;
     __atomic_store_n(&a, 100, 0);
-    printf("atomic: %d\n", a);
+    printf("atomic: %lld\n", a);
     return 0;
 }
